added feature flag toggleFaq to switche between legacy and new FAQ
