+++ react-app/src/features/faq/NewFaq.tsx
@@ -1,0 +1,9 @@
+// import { useEffect, useState } from "react";
+
+export const  NewFaq = (props : any) => {
+  return (
+    <>
+        <h1>New FAQ</h1>
+    </>
+  );
+};
